Fix some GC issues in ToValue
The trait `ToValue` is simplified, and instances are required to
implement a single function `to_value` allocating  an ocaml value (of
type `ocaml_sys::Value`) for a given `T`. It's the responsability of the
caller of `to_value` to either root this value or make it reachable from
a rooted ocaml value before the next allocation point.